requirements.txt
psycopg2==2.9.5
djangorestframework==3.14.0
requests==2.28.2